requirements.txt
streamlit>=1.32.0
transformers>=4.38.0
torch>=2.1.0
torchaudio>=2.1.0
speechbrain>=1.0.0
openai-whisper>=20231117
librosa>=0.10.1
soundfile>=0.12.1
pydub>=0.25.1
pandas>=2.1.0
numpy>=1.26.0
plotly>=5.18.0
scipy>=1.11.0
scikit-learn>=1.3.0
sentencepiece>=0.1.99
protobuf>=4.25.0
accelerate>=0.26.0
datasets>=2.16.0
evaluate>=0.4.1
nltk>=3.8.1
textblob>=0.17.1
pillow>=10.2.0
python-docx>=1.1.0
reportlab>=4.1.0
openpyxl>=3.1.2
tqdm>=4.66.0
huggingface-hub>=0.20.0
